+++ CHANGELOG.md
@@ -17,6 +17,17 @@ Changelog
 * Dropped support for PHP 7.2
 * Added support for PHP 8 Attributes
 
+2.9.2
+-----
+
+### Fixed
+
+* 2.9.1 fixed overwriting flash messages on multiple redirects, but introduced
+  a risk to lose flash messages when redirecting to a path that is outside the
+  firewall or destroys the session.
+  This version hopefully fixes both cases. Existing flash messages in a request
+  cookie are merged with new flash messages from the session.
+
 2.9.1
 -----